requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-generativeai==0.3.2
sentence-transformers>=2.2.2
huggingface-hub>=0.15.0,<0.35.0
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0
pydantic>=2.5.0
numpy>=1.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
torch>=1.9.0
transformers>=4.21.0
